Fix memory profiler type name mapping bug
- Remove incorrect 'NIL' entry from type_names array
- Align type_names array with actual CljType enum indices
- Fix issue where CLJ_SYMBOL was displayed as 'NIL' in profiler output
- Add missing CLJ_CLOSURE type to type_names array
- Ensure correct type display in memory profiling statistics

The bug was caused by misalignment between the type_names array
and the actual CljType enum values, where index 0 was mapped to
'NIL' but should have been 'SYMBOL' (CLJ_SYMBOL = 0).

This fixes the confusing display of 'NIL' in memory profiler output
when nil is correctly represented as NULL in the system.